Refactor editor actions and extend mapping function
A new hook, useEditorItemGestures, is introduced in place of useSpotGroupEditorActions for a more generic approach to bind editor actions that can now work with any map item, not just SpotGroup. Related functions in MapEditorProvider have also been refactored to align with this change. Additionally, mapping function is extended to support texts on the map and `getSpot` function has been enhanced to return more detailed information, including corresponding row and group data, when called.