Update Map Editor tools and improve MapSvg scaling
This commit encompasses a few changes to improve scaling on MapSvg and tidy up the Map Editor tools. The MapSvg changes include implementing new logic to calculate the viewBox and adding window resize listeners to keep track of the container size. In Map Editor tools, some logic was moved to a separate component to reduce complexity and maintain clean code. In addition, areaWidth and areaHeight were added to types.ts.